UIEUS-250-move-actions-to-actions-button get report upload working with deleting ReportUpload
