--- src/components/ReportUpload/index.js
@@ -1,1 +1,0 @@
-export { default } from './ReportUpload';
