+++ src/components/views/UDP.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { get, isEmpty } from 'lodash';
-import { FormattedMessage } from 'react-intl';
+import { FormattedMessage, injectIntl } from 'react-intl';
 
 import { IfPermission, TitleManager } from '@folio/stripes/core';
 import {
@@ -21,6 +21,7 @@ import {
   PaneMenu,
   Row,
 } from '@folio/stripes/components';
+import { Callout } from '@folio/stripes-components';
 import {
   NotesSmartAccordion,
   ViewMetaData,
@@ -32,9 +33,13 @@ import { UDPInfoView } from '../UDPInfo';
 import { HarvestingConfigurationView } from '../HarvestingConfiguration';
 import Statistics from '../Statistics';
 import StartHarvesterModal from '../StartHarvesterModal';
-import ReportUpload from '../ReportUpload';
+// import ReportUpload from '../ReportUpload';
+import CounterUpload from '../ReportUpload/CounterUpload';
+import NonCounterUpload from '../ReportUpload/NonCounterUpload';
 
 import urls from '../../util/urls';
+
+let callout;
 
 class UDP extends React.Component {
   constructor(props) {
@@ -48,7 +53,39 @@ class UDP extends React.Component {
       showCounterUpload: false,
       showNonCounterUpload: false,
     };
+
+    callout = React.createRef();
   }
+
+  handleSuccess = () => {
+    const info = this.props.intl.formatMessage({
+      id: 'ui-erm-usage.report.upload.success',
+    });
+    callout.sendCallout({
+      message: info,
+    });
+    this.setState({
+      showCounterUpload: false,
+      showNonCounterUpload: false,
+    });
+    // this.props.onReloadStatistics();
+    this.reloadStatistics();
+  };
+
+  handleFail = (msg) => {
+    const failText = this.props.intl.formatMessage({
+      id: 'ui-erm-usage.report.upload.failed',
+    });
+    callout.sendCallout({
+      type: 'error',
+      message: `${failText} ${msg}`,
+      timeout: 0,
+    });
+    this.setState({
+      showCounterUpload: false,
+      showNonCounterUpload: false,
+    });
+  };
 
   getInitialAccordionsState = () => {
     return {
@@ -129,13 +166,6 @@ class UDP extends React.Component {
     this.setState({ showNonCounterUpload: true });
   }
 
-  closeReportUploadModal = () => {
-    this.setState({
-      showCounterUpload: false,
-      showNonCounterUpload: false,
-    });
-  }
-
   getActionMenu = () => ({ onToggle }) => {
     const { canEdit, handlers, data } = this.props;
     const usageDataProvider = get(data, 'usageDataProvider', {});
@@ -200,12 +230,6 @@ class UDP extends React.Component {
               <FormattedMessage id="ui-erm-usage.statistics.counter.upload" />
             </Icon>
           </Button>
-          {/* <ReportUpload
-            udpId={providerId}
-            stripes={this.props.stripes}
-            onReloadStatistics={this.reloadStatistics}
-            showCounterUpload={this.state.showCounterUpload}
-          /> */}
         </div>
         <div>
           <Button
@@ -221,23 +245,28 @@ class UDP extends React.Component {
               <FormattedMessage id="ui-erm-usage.statistics.custom.upload" />
             </Icon>
           </Button>
-          {/* <ReportUpload
-            udpId={providerId}
-            stripes={this.props.stripes}
-            onReloadStatistics={this.reloadStatistics}
-            showNonCounterUpload={this.state.showNonCounterUpload}
-          /> */}
         </div>
-        { (this.state.showCounterUpload || this.state.showNonCounterUpload) &&
-        <ReportUpload
-          udpId={providerId}
+        <CounterUpload
+          open={this.state.showCounterUpload}
+          onClose={() => this.setState({ showCounterUpload: false })}
+          onFail={this.handleFail}
+          onSuccess={this.handleSuccess}
           stripes={this.props.stripes}
-          onReloadStatistics={this.reloadStatistics}
-          showCounterUpload={this.state.showCounterUpload}
-          showNonCounterUpload={this.state.showNonCounterUpload}
-          closeReportUploadModal={this.closeReportUploadModal}
+          udpId={providerId}
         />
-        }
+        <NonCounterUpload
+          open={this.state.showNonCounterUpload}
+          onClose={() => this.setState({ showNonCounterUpload: false })}
+          onFail={this.handleFail}
+          onSuccess={this.handleSuccess}
+          stripes={this.props.stripes}
+          udpId={providerId}
+        />
+        <Callout
+          ref={(ref) => {
+            callout = ref;
+          }}
+        />
         {canEdit && (
           <div>
             <Button
@@ -284,10 +313,6 @@ class UDP extends React.Component {
       handler: () => this.handleCloseModal()
     },
   ];
-
-  // handleCloseModal = () => {
-  //   this.setState({ showReportUploadModal: false });
-  // };
 
   renderLoadingPane = () => {
     return (
@@ -370,13 +395,6 @@ class UDP extends React.Component {
                       <FormattedMessage id="ui-erm-usage.udp.harvestingConfiguration" />
                     }
                     id="harvestingAccordion"
-                    // displayWhenOpen={
-                    //   <StartHarvesterModal
-                    //     usageDataProvider={usageDataProvider}
-                    //     isHarvesterExistent={isHarvesterExistent}
-                    //     onReloadUDP={this.reloadUdp}
-                    //   />
-                    // }
                   >
                     <HarvestingConfigurationView
                       usageDataProvider={usageDataProvider}
@@ -398,17 +416,6 @@ class UDP extends React.Component {
                       isStatsLoading={isStatsLoading}
                       handlers={handlers}
                     />
-                  </Accordion>
-                  <Accordion
-                    label={<FormattedMessage id="ui-erm-usage.udp.statsUpload" />}
-                    id="uploadAccordion"
-                  >
-                    {/* <ReportUpload
-                      udpId={providerId}
-                      stripes={stripes}
-                      onReloadStatistics={this.reloadStatistics}
-                      showReportUploadModal={this.state.showReportUploadModal}
-                    /> */}
                   </Accordion>
                   <NotesSmartAccordion
                     id="notesAccordion"
@@ -446,6 +453,7 @@ UDP.propTypes = {
     onClose: PropTypes.func.isRequired,
     onEdit: PropTypes.func,
   }).isRequired,
+  intl: PropTypes.object,
   isHarvesterExistent: PropTypes.bool,
   isLoading: PropTypes.bool.isRequired,
   isStatsLoading: PropTypes.bool.isRequired,
@@ -463,4 +471,4 @@ UDP.propTypes = {
   statsReloadCount: PropTypes.number.isRequired,
 };
 
-export default UDP;
+export default injectIntl(UDP);
